CDC-110: Split code for visual matrix into column and row matrices for faster debugging
