+++ packages/layout/src/visual-matrix/visual-matrix.js
@@ -1,18 +1,13 @@
 import { mergeRecursive, generateGetterSetters } from 'muze-utils';
 import {
     createTree,
-    extraCellsRemover,
     combineMatrices,
-    spaceTakenByColumn,
-    getDistributedHeight,
-    getDistributedWidth,
-    spaceTakenByRow,
     computeLogicalSpace,
     createMatrixEachLevel,
     createMatrixInstances
   } from '../utils';
 import { PROPS } from './props';
-import { COLUMN_ROOT, ROW_ROOT, HEIGHT, WIDTH } from '../enums/constants';
+import { HEIGHT, WIDTH } from '../enums/constants';
 import { defaultConfig } from './default-config';
 
 /**
@@ -41,16 +36,7 @@ export default class VisualMatrix {
         const defCon = Object.assign({}, this.constructor.defaultConfig());
         this.config(mergeRecursive(defCon, config));
 
-        this._breakPointer = this.config().isTransposed ? matrix[0].length :
-            (matrix[0].length > 0 ? matrix[0][0].length : 0);
         this._layoutMatrix = combineMatrices([matrix[0] || [], matrix[1] || []], this.config());
-
-        // Create Tree
-        this._tree = {
-            key: this.config().isTransposed ? COLUMN_ROOT : ROW_ROOT,
-            values: this.createTree()
-        };
-        this._logicalSpace = this.setLogicalSpace();
     }
 
     /**
@@ -158,7 +144,7 @@ export default class VisualMatrix {
         const depth = this.calculateDepth(widthMeasures, heightMeasures);
 
         this.viewableMatrix = this.createViewPortMatrix(depth);
-        this.viewableMeasures = this.redistribute(this.viewableMatrix, width, height);
+        this.viewableMeasures = this.redistributeSpaces(width, height);
         return this;
     }
 
@@ -215,57 +201,21 @@ export default class VisualMatrix {
      * @return {Object} current viewports matrixes with measures
      * @memberof VisualMatrix
      */
-    redistribute (viewableMatrix, width, height) {
+    redistributeSpaces (width, height) {
         let maxHeights = [];
         let maxWidths = [];
-        const {
-            isTransposed
-        } = this.config();
 
-        viewableMatrix.forEach((matrixInst) => {
+        this.viewableMatrix.forEach((matrixInst) => {
             const matrix = matrixInst.matrix;
             const mWidth = 0;
             const mHeight = 0;
             const options = { mWidth, mHeight, matrix, width, height, maxHeights, maxWidths };
-            const maxMeasures = isTransposed ? this.redistributeColumnWise(options) : this.redistributeRowWise(options);
+            const maxMeasures = this.redistributeViewSpaces(options);
             maxWidths = maxMeasures.maxWidths;
             maxHeights = maxMeasures.maxHeights;
         });
 
-        const measurements = viewableMatrix.map((matrixInst, i) => {
-            let heightMeasures;
-            let columnMeasures;
-            const cellDimOptions = { matrixInst, maxWidths, maxHeights, matrixIndex: i };
-            const { heights, widths, rowHeights, columnWidths } = this.getCellDimensions(cellDimOptions);
-
-            if (!isTransposed) {
-                heightMeasures = [height, height];
-                columnMeasures = widths;
-            } else {
-                heightMeasures = heights;
-                columnMeasures = [width, width];
-            }
-            return {
-                rowHeights: {
-                    primary: rowHeights[0],
-                    secondary: rowHeights[1]
-                },
-                columnWidths: {
-                    primary: columnWidths[0],
-                    secondary: columnWidths[1]
-                },
-                height: {
-                    primary: heightMeasures[0],
-                    secondary: heightMeasures[1]
-                },
-                width: {
-                    primary: columnMeasures[0],
-                    secondary: columnMeasures[1]
-                }
-            };
-        });
-
-        return measurements;
+        return this.computeViewableSpaces({ height, width, maxHeights, maxWidths });
     }
 
     /**
@@ -289,36 +239,6 @@ export default class VisualMatrix {
     }
 
     /**
-     *
-     *
-     * @return
-     * @memberof VisualMatrix
-     */
-    removeExtraCells () {
-        const {
-            isTransposed,
-            extraCellLengths
-        } = this.config();
-        const matrix = this._layoutMatrix;
-        const tree = mergeRecursive({}, this.tree());
-        const begCellLen = extraCellLengths[0];
-        const endCellLen = extraCellLengths[1] || Number.NEGATIVE_INFINITY;
-        const layoutMatrix = !isTransposed ? extraCellsRemover(matrix, begCellLen, endCellLen) :
-        matrix.slice(0).map(e => extraCellsRemover(e, begCellLen, endCellLen));
-
-        tree.values = extraCellsRemover(tree.values, begCellLen, endCellLen);
-        if (!isTransposed) {
-            tree.matrix = extraCellsRemover(tree.matrix, begCellLen, endCellLen);
-        } else {
-            tree.matrix = tree.matrix.map(e => extraCellsRemover(e, begCellLen, endCellLen));
-        }
-        return {
-            tree,
-            layoutMatrix
-        };
-    }
-
-    /**
      * Creates the viewport that can be viewed together
      *
      * @param {number} depth maxDepth that can be viewed in the viewport
@@ -329,161 +249,6 @@ export default class VisualMatrix {
         const arr = [];
         createMatrixInstances(arr, depth, this.removeExtraCells(), this);
         return arr;
-    }
-
-    /**
-     * Distibutes the given space row wisely
-     *
-     * @param {Object} options Redistribution information
-     * @memberof VisualMatrix
-     */
-    redistributeRowWise (options) {
-        let cWidths = [];
-        let rHeights = [];
-        let mHeight = 0;
-        const maxMeasures = this.maxMeasures();
-        const {
-            isDistributionEqual,
-            distribution,
-            isTransposed,
-            gutter
-        } = this.config();
-        const { matrix, width, height, maxHeights, maxWidths } = options;
-
-        mHeight = spaceTakenByColumn(matrix, this._lastLevelKey).height;
-
-        const maxWidth = maxMeasures.reduce((t, n) => {
-            t += n;
-            return t;
-        });
-
-        // if (maxWidth > 0) {
-        //     const maxLastRowWidth = Math.min(maxMeasures[maxMeasures.length - 1], width / 2);
-        //     const remainingAvailWidth = width - maxLastRowWidth;
-        //     const remainingWidths = maxWidth - maxLastRowWidth;
-        //     cWidths = maxMeasures.map(space => remainingAvailWidth * (space / remainingWidths));
-        //     cWidths[cWidths.length - 1] = maxLastRowWidth;
-        // }
-        if (maxWidth > 0) {
-            cWidths = maxMeasures.map(space => width * (space / maxWidth));
-        } else {
-            cWidths = maxMeasures.map(() => 0);
-        }
-        rHeights = getDistributedHeight({
-            matrix,
-            cIdx: this._lastLevelKey,
-            height: mHeight,
-            availableHeight: height,
-            isDistributionEqual,
-            distribution,
-            isTransposed,
-            gutter
-        });
-        maxWidths.push(cWidths);
-        maxHeights.push(rHeights);
-        return { maxWidths, maxHeights };
-    }
-
-    /**
-     * Distibutes the given space column wisely
-     *
-     * @param {Object} options Redistribution information
-     * @memberof VisualMatrix
-     */
-    redistributeColumnWise (options) {
-        let rHeights = [];
-        const { matrix, width, height, maxHeights, maxWidths } = options;
-        const borderWidth = this.config().unitMeasures.border;
-        const priority = this.config().priority;
-        const mWidth = spaceTakenByRow(matrix[this._lastLevelKey]).width;
-        const cWidths = getDistributedWidth({
-            row: matrix[this._lastLevelKey],
-            width: mWidth,
-            availableWidth: width
-        }, this.config());
-
-        matrix.forEach((row, rIdx) => row.forEach((col, cIdx) => {
-            const oldLogicalSpace = col.getLogicalSpace().height;
-            col.setAvailableSpace(cWidths[cIdx] - borderWidth, oldLogicalSpace);
-            rHeights[rIdx] = Math.max(rHeights[rIdx] || 0, col.getLogicalSpace().height);
-        }));
-
-        if (maxHeights.length > 0) {
-            rHeights = rHeights.map((e, i) => Math.max(e, maxHeights[0][i]));
-        }
-
-        maxHeights.push(rHeights);
-
-        for (let x = 0; x < maxHeights.length; x++) {
-            maxHeights[x] = rHeights;
-        }
-        maxWidths.push(cWidths);
-        return { maxHeights, maxWidths };
-    }
-
-    /**
-     * Dispatch the calculated cell dimensions to all the cells
-     *
-     * @param {Object} options cell dimension information
-     * @return {Object} row and column heights / widths
-     * @memberof VisualMatrix
-     */
-    getCellDimensions (options) {
-        let indices = [];
-        let unitMeasures = [];
-        let mainMeasures = [];
-        let computedMeasures = [];
-        const {
-            isTransposed,
-            unitMeasures: measures
-        } = this.config();
-        const borderWidth = measures.border;
-        const { matrixInst, maxWidths, maxHeights, matrixIndex } = options;
-        const matrix = matrixInst.matrix;
-        const rowHeights = [[0], [0]];
-        const columnWidths = [[0], [0]];
-        const heights = [0, 0];
-        const widths = [0, 0];
-        const breakPointer = this._breakPointer;
-
-        matrix.forEach((row, rIdx) => {
-            row.forEach((cell, cIdx) => {
-                const colHeight = maxHeights[matrixIndex][rIdx] || 0;
-                const colWidth = maxWidths[matrixIndex][cIdx];
-
-                if (!isTransposed) {
-                    cell.setAvailableSpace(colWidth, colHeight - borderWidth);
-                    indices = [rIdx, cIdx];
-                    unitMeasures = [columnWidths, rowHeights];
-                    mainMeasures = [widths, heights];
-                    computedMeasures = [colWidth, colHeight];
-                } else {
-                    cell.setAvailableSpace(colWidth - borderWidth, colHeight);
-                    indices = [cIdx, rIdx];
-                    unitMeasures = [rowHeights, columnWidths];
-                    mainMeasures = [heights, widths];
-                    computedMeasures = [colHeight, colWidth];
-                }
-
-                if (indices[0] === 0 && indices[1] < breakPointer) {
-                    unitMeasures[0][0][indices[1]] = computedMeasures[0];
-                    mainMeasures[0][0] = (mainMeasures[0][0] || 0) + computedMeasures[0];
-                } else if (indices[0] === 0 && indices[1] >= breakPointer) {
-                    unitMeasures[0][1][indices[1] - breakPointer] = computedMeasures[0];
-                    mainMeasures[0][1] = (mainMeasures[0][1] || 0) + computedMeasures[0];
-                }
-                if (indices[1] === this._lastLevelKey) {
-                    unitMeasures[1][0][indices[0]] = computedMeasures[1];
-                    unitMeasures[1][1][indices[0]] = computedMeasures[1];
-                }
-            });
-        });
-        return {
-            heights,
-            widths,
-            rowHeights,
-            columnWidths
-        };
     }
 }
